src/main/kotlin/eam/edu/co/ingesoft/prestamoLibro/service/UserService.kt
package eam.edu.co.ingesoft.prestamoLibro.service

import eam.edu.co.ingesoft.prestamoLibro.exception.BusinessException
import eam.edu.co.ingesoft.prestamoLibro.model.entities.User
import eam.edu.co.ingesoft.prestamoLibro.repository.UserRepository
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.stereotype.Service
import javax.persistence.EntityManager

@Service
class UserService {
    @Autowired
    lateinit var userRepository: UserRepository

    @Autowired
    lateinit var entityManager: EntityManager

    fun createUser(user: User) {
        val userById = userRepository.find(user.user_identification)

        if (userById != null) {
            throw BusinessException("This person already exists")
        }
        userRepository.create(user)
    }
}